added game over screen and score system for it
