+++ snake/snake.py
@@ -1,168 +1,256 @@
-import pygame, sys, time, random
-
-SIZE = 40
-BACKGROUND = (96, 108, 56)
+import pygame, sys, random
+from pygame.math import Vector2
 
 
-class Apple:
-    def __init__(self, parent_screen):
-        self.apple_image = pygame.image.load("arcade/assets/apple.jpeg")
-        self.parent_screen = parent_screen
-        self.x = SIZE*3
-        self.y = SIZE*3
-
-    def draw(self):
-        self.parent_screen.blit(self.apple_image, (self.x, self.y))
-        pygame.display.update()
-
-    def move(self):
-        self.x = random.randint(0, 24)*SIZE
-        self.y = random.randint(0, 19)*SIZE
-
-
-class Snake:
-    def __init__(self, parent_screen, length):
-        self.length = length
-        self.parent_screen = parent_screen
-        self.block = pygame.image.load("arcade/assets/block.jpeg")
-        self.x = [SIZE] * length
-        self.y = [SIZE] * length
-        self.direction = 'down'
-
-    def draw(self):
-        self.parent_screen.fill(BACKGROUND)
-        for i in range(self.length):
-            self.parent_screen.blit(self.block, (self.x[i], self.y[i]))
-        pygame.display.update()
-
-    def walk(self):
-        for i in range(self.length - 1, 0, -1):
-            self.x[i] = self.x[i - 1]
-            self.y[i] = self.y[i - 1]
-
-        if self.direction == 'left':
-            self.x[0] -= SIZE
-        if self.direction == 'right':
-            self.x[0] += SIZE
-        if self.direction == 'up':
-            self.y[0] -= SIZE
-        if self.direction == 'down':
-            self.y[0] += SIZE
-
-        self.draw()
-
-    def increase_length(self):
-        self.length += 1
-        self.x.append(-1)
-        self.y.append(-1)
-
-    def move_left(self):
-        self.direction = 'left'
-
-    def move_right(self):
-        self.direction = 'right'
-
-    def move_up(self):
-        self.direction = 'up'
-
-    def move_down(self):
-        self.direction = 'down'
-
-
-class Game:
+class SNAKE:
     def __init__(self):
-        pygame.init()
-        self.background_color = (96, 108, 56)
-        self.screen_width = 1000
-        self.screen_height = 800
-        self.clock = pygame.time.Clock()
-        self.screen = pygame.display.set_mode((self.screen_width, self.screen_height))
-        self.snake = Snake(self.screen, 1)
-        self.snake.draw()
-        self.apple = Apple(self.screen)
-        self.apple.draw()
-        pygame.display.set_caption("Snake")
+        self.body = [Vector2(5, 10), Vector2(4, 10), Vector2(3, 10)]
+        self.direction = Vector2(1, 0)
+        self.new_block = False
 
-    def is_collision(self, x1, y1, x2, y2):
-        if x1 >= x2 and x1 < x2 + SIZE:
-            if y1 >= y2 and y1 < y2 + SIZE:
-                return True
+        self.head_up = pygame.image.load('assets/head_up.png')
+        self.head_down = pygame.image.load('assets/head_down.png')
+        self.head_left = pygame.image.load('assets/head_left.png')
+        self.head_right = pygame.image.load('assets/head_right.png')
 
-        return False
-    def display_score(self):
-        font = pygame.font.SysFont('arial', 30)
-        score = font.render(f"Score: {self.snake.length}", True, (255, 255, 255))
-        self.screen.blit(score, (800, 10))
+        self.body_bl = pygame.image.load('assets/body_bl.png')
+        self.body_br = pygame.image.load('assets/body_br.png')
+        self.body_tl = pygame.image.load('assets/body_tl.png')
+        self.body_tr = pygame.image.load('assets/body_tr.png')
 
-    def show_game_over(self):
-        self.screen.fill(BACKGROUND)
-        font = pygame.font.SysFont('arial', 30)
-        line1 = font.render(f"Game is Over: Your score is {self.snake.length}", True, (255, 255, 255))
-        self.screen.blit(line1, (200, 300))
-        line2 = font.render(f"To play again press Enter. To exit press Escape!", True, (255, 255, 255))
-        self.screen.blit(line2, (200, 350))
-        pygame.display.update()
+        self.body_horizontal = pygame.image.load('assets/body_horizontal.png')
+        self.body_vertical = pygame.image.load('assets/body_vertical.png')
 
-    def play(self):
-        self.screen.fill(self.background_color)
-        self.snake.draw()
-        self.snake.walk()
-        self.apple.draw()
-        self.display_score()
-        time.sleep(0.2)
-        self.clock.tick(60)
-        pygame.display.update()
+        self.tail_down = pygame.image.load('assets/tail_down.png')
+        self.tail_left = pygame.image.load('assets/tail_left.png')
+        self.tail_right = pygame.image.load('assets/tail_right.png')
+        self.tail_up = pygame.image.load('assets/tail_up.png')
 
-        # snake colliding with apple
-        if self.is_collision(self.snake.x[0], self.snake.y[0], self.apple.x, self.apple.y):
-            self.snake.increase_length()
-            self.apple.move()
+        self.crunch_sound = pygame.mixer.Sound('assets/Sound_crunch.wav')
 
-        # snake colliding with itself
-        for i in range(1, self.snake.length):
-            if self.is_collision(self.snake.x[0], self.snake.y[0], self.snake.x[i], self.snake.y[i]):
-                raise Exception("Game Over")
+    def draw_snake(self):
+        self.update_head_graphics()
+        self.update_tail_graphics()
+
+        for index, block in enumerate(self.body):
+            x_pos = int(block.x * cell_size)
+            y_pos = int(block.y * cell_size)
+            block_rect = pygame.Rect(x_pos, y_pos, cell_size, cell_size)
+
+            if index == 0:
+                screen.blit(self.head, block_rect)
+            elif index == len(self.body) - 1:
+                screen.blit(self.tail, block_rect)
+            else:
+                previous_block = self.body[index + 1] - block
+                next_block = self.body[index -1] - block
+                if previous_block.x == next_block.x:
+                    screen.blit(self.body_vertical, block_rect)
+                elif previous_block.y == next_block.y:
+                    screen.blit(self.body_horizontal, block_rect)
+                else:
+                    if previous_block.x == -1 and next_block.y == -1 or previous_block.y == -1 and next_block.x == -1:
+                        screen.blit(self.body_tl, block_rect)
+                    elif previous_block.x == -1 and next_block.y == 1 or previous_block.y == 1 and next_block.x == -1:
+                        screen.blit(self.body_bl, block_rect)
+                    elif previous_block.x == 1 and next_block.y == -1 or previous_block.y == -1 and next_block.x == 1:
+                        screen.blit(self.body_tr, block_rect)
+                    elif previous_block.x == 1 and next_block.y == 1 or previous_block.y == 1 and next_block.x == 1:
+                        screen.blit(self.body_br, block_rect)
+
+    def update_head_graphics(self):
+        head_relation = self.body[1] - self.body[0]
+        if head_relation == Vector2(1, 0):
+            self.head = self.head_left
+        elif head_relation == Vector2(-1, 0):
+            self.head = self.head_right
+        elif head_relation == Vector2(0, 1):
+            self.head = self.head_up
+        elif head_relation == Vector2(0, -1):
+            self.head = self.head_down
 
 
+    def update_tail_graphics(self):
+        tail_relation = self.body[-2] - self.body[-1]
+        if tail_relation == Vector2(1, 0):
+            self.tail = self.tail_left
+        elif tail_relation == Vector2(-1, 0):
+            self.tail = self.tail_right
+        elif tail_relation == Vector2(0, 1):
+            self.tail = self.tail_up
+        elif tail_relation == Vector2(0, -1):
+            self.tail = self.tail_down
+
+    def move_snake(self):
+        if self.new_block == True:
+            body_copy = self.body[:]
+            body_copy.insert(0, body_copy[0] + self.direction)
+            self.body = body_copy[:]
+            self.new_block = False
+        else:
+            body_copy = self.body[:-1]
+            body_copy.insert(0, body_copy[0] + self.direction)
+            self.body = body_copy[:]
+
+    def add_block(self):
+        self.new_block = True
+
+    def play_crunch_sound(self):
+        self.crunch_sound.play()
 
     def reset(self):
-        self.snake = Snake(self.screen, 1)
-        self.apple = Apple(self.screen)
-
-    def run(self):
-        running = True
-        pause = False
-        while running:
-            # Handling input
-            for event in pygame.event.get():
-                if event.type == pygame.KEYDOWN:
-                    if event.key == pygame.K_ESCAPE:
-                        running = False
-                    if event.key == pygame.K_RETURN:
-                        pause = False
-                    if not pause:
-                        if event.key == pygame.K_UP:
-                            self.snake.move_up()
-                        if event.key == pygame.K_DOWN:
-                            self.snake.move_down()
-                        if event.key == pygame.K_LEFT:
-                            self.snake.move_left()
-                        if event.key == pygame.K_RIGHT:
-                            self.snake.move_right()
-                elif event.type == pygame.QUIT:
-                    pygame.quit()
-                    sys.exit()
-
-            try:
-                if not pause:
-                    self.play()
-            except Exception as e:
-                self.show_game_over()
-                pause = True
-                self.reset()
+        self.body = [Vector2(5, 10), Vector2(4, 10), Vector2(3, 10)]
 
 
-# Updating the window
-if __name__ == '__main__':
-    game = Game()
-    game.run()
+class FRUIT:
+    def __init__(self):
+        self.randomize()
+
+    def draw_fruit(self):
+        fruit_rect = pygame.Rect(self.pos.x * cell_size, self.pos.y * cell_size, cell_size, cell_size)
+        screen.blit(apple, fruit_rect)
+
+    def randomize(self):
+        self.x = random.randint(0, cell_number - 1)
+        self.y = random.randint(0, cell_number - 1)
+        self.pos = Vector2(self.x, self.y)
+
+
+class MAIN:
+    def __init__(self):
+        self.snake = SNAKE()
+        self.fruit = FRUIT()
+        self.current_score = 0
+
+    def update(self):
+        self.snake.move_snake()
+        self.check_collision()
+        self.check_fail()
+
+    def draw_elements(self):
+        self.draw_grass()
+        self.fruit.draw_fruit()
+        self.snake.draw_snake()
+        self.draw_score()
+
+    def check_collision(self):
+        if self.fruit.pos == self.snake.body[0]:
+            self.fruit.randomize()
+            self.snake.add_block()
+            self.snake.play_crunch_sound()
+            self.current_score += 1
+
+        for block in self.snake.body[1:]:
+            if block == self.fruit.pos:
+                self.fruit.randomize()
+
+    def check_fail(self):
+        if not 0 <= self.snake.body[0].x < cell_number or not 0 <= self.snake.body[0].y < cell_number:
+            # self.show_game_over()
+            self.game_over()
+
+        for block in self.snake.body[1:]:
+            if block == self.snake.body[0]:
+                # self.show_game_over()
+                self.game_over()
+
+    def show_game_over(self):
+        screen.fill(BACKGROUND)
+        line1 = game_font.render(f"Game is Over: Your score is {self.current_score}", True, (255, 255, 255))
+        screen.blit(line1, (200, 300))
+        line2 = game_font.render(f"To play again press Enter. To exit press Escape!", True, (255, 255, 255))
+        screen.blit(line2, (200, 350))
+        pygame.display.update()
+
+    def game_over(self):
+        global pause
+        pause = True
+        self.snake.reset()
+
+
+
+    def draw_grass(self):
+        grass_color = (167, 209, 61)
+
+        for row in range(cell_number):
+            if row % 2 == 0:
+                for col in range(cell_number):
+                    if col % 2 == 0:
+                        grass_rect = pygame.Rect(col * cell_size, row * cell_size, cell_size, cell_size)
+                        pygame.draw.rect(screen, grass_color, grass_rect)
+            else:
+                for col in range(cell_number):
+                    if col % 2 != 0:
+                        grass_rect = pygame.Rect(col * cell_size, row * cell_size, cell_size, cell_size)
+                        pygame.draw.rect(screen, grass_color, grass_rect)
+
+    def draw_score(self):
+        score_text = str(len(self.snake.body) - 3)
+        score_surface = game_font.render(score_text, True, (56, 74, 12))
+        score_x = int(cell_size * cell_number - 60)
+        score_y = int(cell_size * cell_number - 40)
+        score_rect = score_surface.get_rect(center=(score_x, score_y))
+        apple_rect = apple.get_rect(midright=(score_rect.left,score_rect.centery))
+        bg_rect = pygame.Rect(apple_rect.left, apple_rect.top, apple_rect.width + score_rect.width + 6, apple_rect.height)
+
+        pygame.draw.rect(screen, (167, 209, 61),bg_rect)
+        screen.blit(score_surface, score_rect)
+        screen.blit(apple, apple_rect)
+        pygame.draw.rect(screen, (56, 74, 12), bg_rect, 2)
+
+
+pygame.mixer.pre_init(44100, -16, 2, 512)
+pygame.init()
+cell_size = 40
+cell_number = 20
+BACKGROUND = (96, 108, 56)
+screen = pygame.display.set_mode((cell_number * cell_size, cell_number * cell_size))
+clock = pygame.time.Clock()
+apple = pygame.image.load('assets/apple.png')
+game_font = pygame.font.Font('assets/PoetsenOne-Regular.ttf', 25)
+pause = False
+running = True
+main_game = MAIN()
+current_score = 0
+
+SCREEN_UPDATE = pygame.USEREVENT
+pygame.time.set_timer(SCREEN_UPDATE, 150)
+
+while running:
+    for event in pygame.event.get():
+        if event.type == pygame.QUIT:
+            pygame.quit()
+            sys.exit()
+        if pause is False:
+            if event.type == SCREEN_UPDATE:
+                main_game.update()
+            if event.type == pygame.KEYDOWN:
+                if event.key == pygame.K_ESCAPE:
+                    running = False
+                if event.key == pygame.K_RETURN:
+                    running = True
+                if event.key == pygame.K_UP:
+                    if main_game.snake.direction.y != 1:
+                        main_game.snake.direction = Vector2(0, -1)
+                if event.key == pygame.K_DOWN:
+                    if main_game.snake.direction.y != -1:
+                        main_game.snake.direction = Vector2(0, 1)
+                if event.key == pygame.K_LEFT:
+                    if main_game.snake.direction.x != 1:
+                        main_game.snake.direction = Vector2(-1, 0)
+                if event.key == pygame.K_RIGHT:
+                    if main_game.snake.direction.x != -1:
+                        main_game.snake.direction = Vector2(1, 0)
+
+            screen.fill((175, 215, 70))
+            main_game.draw_elements()
+            pygame.display.update()
+            clock.tick(60)
+        elif pause is True:
+            main_game.show_game_over()
+            if event.type == pygame.KEYDOWN:
+                if event.key == pygame.K_RETURN:
+                    # running = True
+                    pause = False
+                if event.key == pygame.K_ESCAPE:
+                    running = False
