+++ tests/test_snake.py
@@ -1,0 +1,2 @@
+import pytest
+
